Updates after Semrush testing

+++ src/main/java/com/searchify/suggestion/api/response/TopSubfolderResponse.java
@@ -25,5 +25,5 @@ public class TopSubfolderResponse {
     private YearMonth displayDate;
 
     private Double trafficShare;
-    private Integer uniquePageViews;
+    private Long uniquePageViews;
 }
